# docs/openapi_3.yaml
openapi: 3.0.0
info:
  title: Simple Calc API
  description: This is a simple Calc API
  contact:
    email: eric@erb-tech.com
  version: 1.0.0
servers:
  - url: http://localhost:34349
    description: http url
paths:
  /add:
    post:
      summary: adds two numbers
      description: Adds two operands and returns the result
      operationId: add
      requestBody:
        description: Adds numbers
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Operands'
      responses:
        "200":
          description: 200OK
        "400":
          description: "bad request"
        "500":
          description: server error
  /sub:
    post:
      summary: Subtracts first number from the second
      description: Subtracts two numbers
      operationId: Sub
      requestBody:
        description: Subtracts numbers
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Operands'
      responses:
        "200":
          description: 200OK
        "400":
          description: "bad request"
        "500":
          description: server error
  /mult:
    post:
      summary: Multiplies operands
      description: multiplies two numbers
      operationId: mult
      requestBody:
        description: Multiplies numbers
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Operands'
      responses:
        "200":
          description: 200OK
        "400":
          description: "bad request"
        "500":
          description: server error
  /div:
    post:
      summary: Divides operands
      description: divides two numbers
      operationId: div
      requestBody:
        description: Divides numbers
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Operands'
      responses:
        "200":
          description: 200OK
        "400":
          description: "bad Request"
        "500":
          description: server error
  /sqrt:
    post:
      summary: Square root of a number
      description: Square root of a number
      operationId: sqrt
      requestBody:
        description: Square root of a number
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/sqrt'
      responses:
        "200":
          description: 200OK
        "400":
          description: "bad request"
        "500":
          description: server error
  /login:
    post:
      summary: User login
      description: Logs in a user (username, password)
      operationId: login
      requestBody:
        description: get username and password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login'
      responses:
        "200":
          description: 200OK
        "400":
          description: "bad request"
        "500":
          description: server error
components:
  schemas:
    Operands:
      required:
        - left
        - right
      type: object
      properties:
        left:
          type: number
          example: 1
        right:
          type: number
          example: 48030
    sqrt:
      required:
        - sqrt
      type: object
      properties:
        sqrt:
          type: number
          example: 100
    login:
      required:
        - username
        - password
      type: object
      properties:
        username:
            type: string
            example: "testuser@user.com"
        password:
            type: string
            example: "myPass"